First attempt at DVE code Refactor
