+++ nautobot/core/settings.py
@@ -575,11 +575,11 @@ INSTALLED_APPS = [
     "nautobot.dcim",
     "nautobot.ipam",
     "nautobot.extras",
+    "nautobot.nautobot_data_validation_engine",
     "nautobot.tenancy",
     "nautobot.users",
     "nautobot.virtualization",
     "nautobot.wireless",
-    "nautobot.nautobot_data_validation_engine",
     "drf_spectacular",
     "drf_spectacular_sidecar",
     "graphene_django",
