+++ nautobot/nautobot_data_validation_engine/forms.py
@@ -3,10 +3,7 @@
 from django import forms
 from django.contrib.contenttypes.models import ContentType
 
-try:
-    from nautobot.apps.constants import CHARFIELD_MAX_LENGTH
-except ImportError:
-    CHARFIELD_MAX_LENGTH = 255
+from nautobot.apps.constants import CHARFIELD_MAX_LENGTH
 from nautobot.core.forms import (
     BootstrapMixin,
     BulkEditNullBooleanSelect,
