+++ nautobot/nautobot_data_validation_engine/models.py
@@ -6,12 +6,8 @@ from django.contrib.contenttypes.fields import GenericForeignKey
 from django.contrib.contenttypes.models import ContentType
 from django.core.validators import MinValueValidator, ValidationError
 from django.db import models
-from django.shortcuts import reverse
 
-try:
-    from nautobot.apps.constants import CHARFIELD_MAX_LENGTH
-except ImportError:
-    CHARFIELD_MAX_LENGTH = 255
+from nautobot.core.constants import CHARFIELD_MAX_LENGTH
 from nautobot.core.models.generics import PrimaryModel
 from nautobot.core.models.querysets import RestrictedQuerySet
 from nautobot.extras.utils import extras_features, FeatureQuery
@@ -42,13 +38,19 @@ class ValidationRuleManager(RestrictedQuerySet):
 class ValidationRule(PrimaryModel):
     """Base model for all validation engine rule models."""
 
-    name = models.CharField(max_length=100, unique=True)
+    name = models.CharField(max_length=CHARFIELD_MAX_LENGTH, unique=True)
+    field = models.CharField(
+        max_length=50,
+    )
     content_type = models.ForeignKey(
         to=ContentType, on_delete=models.CASCADE, limit_choices_to=FeatureQuery("custom_validators")
     )
     enabled = models.BooleanField(default=True)
     error_message = models.CharField(
-        max_length=255, blank=True, default="", help_text="Optional error message to display when validation fails."
+        max_length=CHARFIELD_MAX_LENGTH,
+        blank=True,
+        default="",
+        help_text="Optional error message to display when validation fails.",
     )
 
     objects = ValidationRuleManager.as_manager()
@@ -69,15 +71,11 @@ class ValidationRule(PrimaryModel):
     "export_templates",
     "graphql",
     "relationships",
-    "statuses",
     "webhooks",
 )
 class RegularExpressionValidationRule(ValidationRule):
     """A type of validation rule that applies a regular expression to a given model field."""
 
-    field = models.CharField(
-        max_length=50,
-    )
     regular_expression = models.TextField()
     context_processing = models.BooleanField(
         default=False,
@@ -91,10 +89,6 @@ class RegularExpressionValidationRule(ValidationRule):
 
         ordering = ("name",)
         unique_together = [["content_type", "field"]]
-
-    def get_absolute_url(self, api=False):
-        """Absolute url for the instance."""
-        return reverse("nautobot_data_validation_engine:regularexpressionvalidationrule", args=[self.id])
 
     def clean(self):
         """Ensure field is valid for the model and has not been blacklisted."""
@@ -140,15 +134,11 @@ class RegularExpressionValidationRule(ValidationRule):
     "export_templates",
     "graphql",
     "relationships",
-    "statuses",
     "webhooks",
 )
 class MinMaxValidationRule(ValidationRule):
     """A type of validation rule that applies min/max constraints to a given numeric model field."""
 
-    field = models.CharField(
-        max_length=50,
-    )
     min = models.FloatField(
         null=True, blank=True, help_text="When set, apply a minimum value contraint to the value of the model field."
     )
@@ -164,10 +154,6 @@ class MinMaxValidationRule(ValidationRule):
         ordering = ("name",)
         unique_together = [["content_type", "field"]]
 
-    def get_absolute_url(self, api=False):
-        """Absolute url for the instance."""
-        return reverse("nautobot_data_validation_engine:minmaxvalidationrule", args=[self.id])
-
     def clean(self):
         """Ensure field is valid for the model and has not been blacklisted."""
         if self.field not in [f.name for f in self.content_type.model_class()._meta.get_fields()]:
@@ -177,21 +163,21 @@ class MinMaxValidationRule(ValidationRule):
                 }
             )
 
-        whitelisted_field_types = (
+        allowed_field_types = (
             models.DecimalField,
             models.FloatField,
             models.IntegerField,
         )
 
-        blacklisted_field_types = (
+        excluded_field_types = (
             models.AutoField,
             models.BigAutoField,
         )
 
         model_field = self.content_type.model_class()._meta.get_field(self.field)
 
-        if not isinstance(model_field, whitelisted_field_types) or (
-            self.field.startswith("_") or not model_field.editable or isinstance(model_field, blacklisted_field_types)
+        if not isinstance(model_field, allowed_field_types) or (
+            self.field.startswith("_") or not model_field.editable or isinstance(model_field, excluded_field_types)
         ):
             raise ValidationError({"field": "This field's type does not support min/max validation."})
 
@@ -213,15 +199,10 @@ class MinMaxValidationRule(ValidationRule):
     "export_templates",
     "graphql",
     "relationships",
-    "statuses",
     "webhooks",
 )
 class RequiredValidationRule(ValidationRule):
     """A type of validation rule that applies a required constraint to a given model field."""
-
-    field = models.CharField(
-        max_length=50,
-    )
 
     clone_fields = ["enabled", "content_type", "error_message"]
 
@@ -230,10 +211,6 @@ class RequiredValidationRule(ValidationRule):
 
         ordering = ("name",)
         unique_together = [["content_type", "field"]]
-
-    def get_absolute_url(self, api=False):
-        """Absolute url for the instance."""
-        return reverse("nautobot_data_validation_engine:requiredvalidationrule", args=[self.id])
 
     def clean(self):
         """Ensure field is valid for the model and has not been blacklisted."""
@@ -269,7 +246,6 @@ class RequiredValidationRule(ValidationRule):
     "export_templates",
     "graphql",
     "relationships",
-    "statuses",
     "webhooks",
 )
 class UniqueValidationRule(ValidationRule):
@@ -279,9 +255,6 @@ class UniqueValidationRule(ValidationRule):
     Optionally specify the max number of similar values for the field accross all model instances. Default of 1.
     """
 
-    field = models.CharField(
-        max_length=50,
-    )
     max_instances = models.PositiveIntegerField(default=1, validators=[MinValueValidator(1)])
 
     clone_fields = ["enabled", "content_type", "max_instances", "error_message"]
@@ -291,10 +264,6 @@ class UniqueValidationRule(ValidationRule):
 
         ordering = ("name",)
         unique_together = [["content_type", "field"]]
-
-    def get_absolute_url(self, api=False):
-        """Absolute url for the instance."""
-        return reverse("nautobot_data_validation_engine:uniquevalidationrule", args=[self.id])
 
     def clean(self):
         """Ensure field is valid for the model and has not been blacklisted."""
@@ -349,7 +318,3 @@ class DataCompliance(PrimaryModel):
     def __str__(self):
         """Return a string representation of this DataCompliance object."""
         return f"{self.compliance_class_name}: {self.validated_attribute} compliance for {self.validated_object}"
-
-    def get_absolute_url(self, api=False):
-        """Return the absolute URL to this Audit object."""
-        return reverse("nautobot_data_validation_engine:datacompliance", args=[self.pk])
