+++ nautobot/nautobot_data_validation_engine/tests/test_views.py
@@ -1,11 +1,9 @@
 """Unit tests for nautobot_data_validation_engine views."""
 
-from unittest import skipIf
 from unittest.mock import MagicMock, patch
 
 from django.contrib.contenttypes.models import ContentType
 from django.http.request import QueryDict
-from packaging import version
 
 from nautobot.core.testing import TestCase, ViewTestCases
 from nautobot.dcim.models import Device, Location, LocationType, PowerFeed
@@ -21,28 +19,11 @@ from nautobot.nautobot_data_validation_engine.tables import DataComplianceTableT
 from nautobot.nautobot_data_validation_engine.tests.test_data_compliance_rules import TestFailedDataComplianceRule
 from nautobot.nautobot_data_validation_engine.views import DataComplianceObjectView
 
-try:
-    from importlib import metadata
-except ImportError:
-    # Running on pre-3.8 Python; use importlib-metadata package
-    import importlib_metadata as metadata
-
-_NAUTOBOT_VERSION = version.parse(metadata.version("nautobot"))
-# Related to this issue: https://github.com/nautobot/nautobot/issues/2948
-_FAILING_OBJECT_LIST_NAUTOBOT_VERSIONS = [version.parse("1.5.2"), version.parse("1.5.3"), version.parse("1.5.4")]
-
 
 class RegularExpressionValidationRuleTestCase(ViewTestCases.PrimaryObjectViewTestCase):
     """View test cases for the RegularExpressionValidationRule model."""
 
     model = RegularExpressionValidationRule
-
-    @skipIf(
-        _NAUTOBOT_VERSION in _FAILING_OBJECT_LIST_NAUTOBOT_VERSIONS,
-        f"Skip test in Nautobot version {_NAUTOBOT_VERSION} due to Nautobot issue #2948",
-    )
-    def test_list_objects_with_permission(self):
-        super().test_list_objects_with_permission()
 
     @classmethod
     def setUpTestData(cls):
@@ -93,13 +74,6 @@ class MinMaxValidationRuleTestCase(ViewTestCases.PrimaryObjectViewTestCase):
     """View test cases for the MinMaxValidationRule model."""
 
     model = MinMaxValidationRule
-
-    @skipIf(
-        _NAUTOBOT_VERSION in _FAILING_OBJECT_LIST_NAUTOBOT_VERSIONS,
-        f"Skip test in Nautobot version {_NAUTOBOT_VERSION} due to Nautobot issue #2948",
-    )
-    def test_list_objects_with_permission(self):
-        super().test_list_objects_with_permission()
 
     @classmethod
     def setUpTestData(cls):
@@ -153,13 +127,6 @@ class RequiredValidationRuleTestCase(ViewTestCases.PrimaryObjectViewTestCase):
 
     model = RequiredValidationRule
 
-    @skipIf(
-        _NAUTOBOT_VERSION in _FAILING_OBJECT_LIST_NAUTOBOT_VERSIONS,
-        f"Skip test in Nautobot version {_NAUTOBOT_VERSION} due to Nautobot issue #2948",
-    )
-    def test_list_objects_with_permission(self):
-        super().test_list_objects_with_permission()
-
     @classmethod
     def setUpTestData(cls):
         """
@@ -204,13 +171,6 @@ class UniqueValidationRuleTestCase(ViewTestCases.PrimaryObjectViewTestCase):
     """View test cases for the UniqueValidationRule model."""
 
     model = UniqueValidationRule
-
-    @skipIf(
-        _NAUTOBOT_VERSION in _FAILING_OBJECT_LIST_NAUTOBOT_VERSIONS,
-        f"Skip test in Nautobot version {_NAUTOBOT_VERSION} due to Nautobot issue #2948",
-    )
-    def test_list_objects_with_permission(self):
-        super().test_list_objects_with_permission()
 
     @classmethod
     def setUpTestData(cls):
@@ -266,13 +226,6 @@ class DataComplianceTestCase(
     """Test cases for DataCompliance Viewset."""
 
     model = DataCompliance
-
-    @skipIf(
-        _NAUTOBOT_VERSION in _FAILING_OBJECT_LIST_NAUTOBOT_VERSIONS,
-        f"Skip test in Nautobot version {_NAUTOBOT_VERSION} due to Nautobot issue #2948",
-    )
-    def test_list_objects_with_permission(self):
-        super().test_list_objects_with_permission()
 
     @classmethod
     def setUpTestData(cls):
